+++ src/models/embeddings.py
@@ -4,14 +4,10 @@ Determine the similarity between video tags and predefined stereotypes by levera
 
 import json
 import logging
-import os
 from typing import Dict, List
 
-import matplotlib.pyplot as plt
-import networkx as nx
 import nltk
 import pandas as pd
-import seaborn as sns
 from nltk.corpus import stopwords
 from nltk.tokenize import word_tokenize
 from sentence_transformers import SentenceTransformer
@@ -21,12 +17,6 @@ from sklearn.metrics.pairwise import cosine_similarity
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
 )
-
-
-def ensure_directory_exists(directory: str):
-    """Ensure that a directory exists; if not, create it."""
-    if not os.path.exists(directory):
-        os.makedirs(directory)
 
 
 def download_nltk_data():
@@ -112,60 +102,3 @@ def calculate_similarity(
     stereotypes_keys, stereotypes_values = zip(*stereotypes_emb.items())
     sim_matrix = cosine_similarity(tags_values, stereotypes_values)
     return pd.DataFrame(sim_matrix, index=tags_keys, columns=stereotypes_keys)
-
-
-def plot_heatmap(df: pd.DataFrame, file_name: str, model_name: str):
-    ensure_directory_exists("plots")
-    plt.figure(figsize=(10, 8))
-    sns.heatmap(df, annot=True, cmap="coolwarm", fmt=".2f")
-    plt.title("Cosine Similarity between Videos and Stereotypes")
-    plt.xlabel("Stereotypes")
-    plt.ylabel("Videos")
-    plt.savefig(f"plots/{file_name}_{model_name}")
-    plt.show()
-
-
-def create_network_graph(
-    tags: Dict[str, str],
-    stereotypes: Dict[str, str],
-    df: pd.DataFrame,
-    file_name: str,
-    model_name: str,
-):
-    """
-    Create and plot a network graph based on similarity scores between tags and stereotypes.
-
-    Parameters:
-    - tags (Dict[str, str]): A dictionary where keys are tag names and values are their descriptions.
-    - stereotypes (Dict[str, str]): A dictionary where keys are stereotype names and values are their descriptions.
-    - sim (pd.DataFrame): A DataFrame containing the cosine similarity scores between tags and stereotypes.
-    - file_name (str): The name of the file to save the network graph plot.
-
-    The function generates a network graph where nodes represent tags and stereotypes, and edges represent the similarity scores between them. Nodes are grouped by type (tag or stereotype), and edge thickness is proportional to the similarity score.
-    """
-    ensure_directory_exists("plots")
-    G = nx.Graph()
-    G.add_nodes_from(tags.keys(), bipartite=0)
-    G.add_nodes_from(stereotypes.keys(), bipartite=1)
-    for i, tag in enumerate(tags.keys()):
-        for j, stereotype in enumerate(stereotypes.keys()):
-            weight = df.iloc[i, j]
-            if weight > 0.3:
-                G.add_edge(tag, stereotype, weight=weight)
-    pos = nx.spring_layout(G, k=0.5, iterations=50)
-    plt.figure(figsize=(12, 8))
-    edges = G.edges(data=True)
-    weights = [edge[2]["weight"] for edge in edges]
-    nx.draw(
-        G,
-        pos,
-        with_labels=True,
-        node_size=3000,
-        node_color="skyblue",
-        font_size=10,
-        font_weight="bold",
-    )
-    nx.draw_networkx_edges(G, pos, edge_color=weights, edge_cmap=plt.cm.Blues, width=2)
-    plt.title("Network Graph of Cosine Similarity between Tags and Stereotypes")
-    plt.savefig(f"plots/{file_name}_{model_name}")
-    plt.show()
