+++ src/models/stereotypes.json
@@ -1,0 +1,9 @@
+{
+"SubWhW" : "White women as submissive, sexually innocent or subservient; a more specific stereotype that falls within this category is that of the Schoolgirl, another that fits within this frame is the Slut stereotype used to denote women who do want to have sex, the Mom stereotype may also fall within this category", 
+"AggresBlM" : "Black men as overly aggressive and endowed with a heightened libido and large genitalia; specific stereotypes that fall within this category are those of the Thug, Gang and BBC (Big Black Cock)",
+"LargeBlW" : "Black women also as endowed with a heightened libido and large bottoms; a specific stereotype that falls within this category is that of the Jezebel",
+"AggresM" : "Men as aggressive and dominant", 
+"AsianW" : "Asian women as exotic or innocent; specific stereotypes that fall within this category are those of the Dragon Lady and Lotus Flower",
+"Taboo" : "Taboo subjects, like incest narratives; specific stereotypical figures that fall within this category are the stepmom and step brother/sister",
+"LargeBreastWhW" : "White women with very large breasts and small waists"
+}
